+++ changelog.md
@@ -56,6 +56,7 @@
 * Base price in product list ignored PriceDisplayType (catalog settings) and possibly displayed the wrong base price info
 * Private messages: Fixes "No route in the route table matches the supplied values"
 * Payone: Hash string incorrect for frontend API payments where the order has more than 9 products
+* Export mail notification: Download link not working if SSL is enabled
 
 
 ## SmartStore.NET 2.5